# project/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Leprecoin API",
    "description": "API for personal finance management application",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/auth/register": {
      "post": {
        "tags": ["auth"],
        "summary": "Register",
        "operationId": "register_api_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UserRegister" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UserResponse" }
              }
            }
          },
          "400": { "description": "Email already registered" },
          "422": { "description": "Validation Error" }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": ["auth"],
        "summary": "Login",
        "operationId": "login_api_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UserLogin" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UserResponse" }
              }
            }
          },
          "401": { "description": "Invalid credentials" },
          "422": { "description": "Validation Error" }
        }
      }
    },
...
    "/api/auth/me": {
      "get": {
        "tags": ["auth"],
        "summary": "Get Me",
        "operationId": "get_me_api_auth_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UserResponse" }
              }
            }
          },
          "401": { "description": "Not authenticated" }
        },
        "security": [{ "HTTPBearer": [] }]
      }
    },
...
    "/api/incomes/{income_id}": {
      "put": {
        "tags": ["expenses"],
        "summary": "Update Income",
        "operationId": "update_income_api_incomes__income_id__put",
        "parameters": [
          {
            "name": "income_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/IncomeUpdate" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Income" }
              }
            }
          },
          "404": { "description": "Income not found" }
        },
        "security": [{ "HTTPBearer": [] }]
      },
      "delete": {
        "tags": ["expenses"],
        "summary": "Delete Income",
        "operationId": "delete_income_api_incomes__income_id__delete",
        "parameters": [
          {
            "name": "income_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" }
                  }
                }
              }
            }
          },
          "404": { "description": "Income not found" }
        },
        "security": [{ "HTTPBearer": [] }]
      }
    },
    "/api/fixed-expenses/{expense_id}": {
      "put": {
        "tags": ["expenses"],
        "summary": "Update Fixed Expense",
        "operationId": "update_fixed_expense_api_fixed_expenses__expense_id__put",
        "parameters": [
          {
            "name": "expense_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/FixedExpenseUpdate" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/FixedExpense" }
              }
            }
          },
          "404": { "description": "Fixed expense not found" }
        },
        "security": [{ "HTTPBearer": [] }]
      },
      "delete": {
        "tags": ["expenses"],
        "summary": "Delete Fixed Expense",
        "operationId": "delete_fixed_expense_api_fixed_expenses__expense_id__delete",
        "parameters": [
          {
            "name": "expense_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" }
                  }
                }
              }
            }
          },
          "404": { "description": "Fixed expense not found" }
        },
        "security": [{ "HTTPBearer": [] }]
      }
    },
    "/api/daily-expenses/{expense_id}": {
      "put": {
        "tags": ["expenses"],
        "summary": "Update Daily Expense",
        "operationId": "update_daily_expense_api_daily_expenses__expense_id__put",
        "parameters": [
          {
            "name": "expense_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/DailyExpenseUpdate" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DailyExpense" }
              }
            }
          },
          "404": { "description": "Daily expense not found" }
        },
        "security": [{ "HTTPBearer": [] }]
      },
      "delete": {
        "tags": ["expenses"],
        "summary": "Delete Daily Expense",
        "operationId": "delete_daily_expense_api_daily_expenses__expense_id__delete",
        "parameters": [
          {
            "name": "expense_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" }
                  }
                }
              }
            }
          },
          "404": { "description": "Daily expense not found" }
        },
        "security": [{ "HTTPBearer": [] }]
      }
    },
    "/api/categories": {
      "get": {
        "tags": ["categories"],
        "summary": "Get Categories",
        "operationId": "get_categories_api_categories_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Category" }
                }
              }
            }
          }
        },
        "security": [{ "HTTPBearer": [] }]
      }
    },
    "/api/detect-category": {
      "post": {
        "tags": ["categories"],
        "summary": "Detect Category",
        "operationId": "detect_category_api_detect_category_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/DetectCategoryRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DetectCategoryResponse" }
              }
            }
          }
        },
        "security": [{ "HTTPBearer": [] }]
      }
    }
  },
  "components": {
    "schemas": {
      "UserRegister": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": { "type": "string", "format": "email" },
          "password": { "type": "string" }
        }
      },
      "UserLogin": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": { "type": "string", "format": "email" },
          "password": { "type": "string" }
        }
      },
      "UserResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "email": { "type": "string" },
          "is_active": { "type": "boolean" }
        }
      },
...
      "IncomeUpdate": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "amount": { "type": "number" }
        }
      },
      "Income": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "month_id": { "type": "integer" },
          "name": { "type": "string" },
          "amount": { "type": "number" }
        }
      },
...
      "FixedExpenseUpdate": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "amount": { "type": "number" }
        }
      },
      "FixedExpense": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "month_id": { "type": "integer" },
          "name": { "type": "string" },
          "amount": { "type": "number" }
        }
      },
...
      "DailyExpenseUpdate": {
        "type": "object",
        "properties": {
          "date": { "type": "string", "format": "date" },
          "description": { "type": "string" },
          "amount": { "type": "number" },
          "category": { "type": "string" }
        }
      },
      "DailyExpense": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "month_id": { "type": "integer" },
          "date": { "type": "string", "format": "date" },
          "description": { "type": "string" },
          "amount": { "type": "number" },
          "category": { "type": "string" }
        }
      },
...
      "Category": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "name": { "type": "string" },
          "keywords": { "type": "string" }
        }
      },
      "DetectCategoryRequest": {
        "type": "object",
        "required": ["description"],
        "properties": {
          "description": { "type": "string" }
        }
      },
      "DetectCategoryResponse": {
        "type": "object",
        "properties": {
          "category": { "type": "string" }
        }
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}